backend/pyproject.toml
[project]
name = "voyagerhack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.4.6",
    "aiohttp==3.11.12",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anthropic==0.45.2",
    "anyio==4.7.0",
    "asttokens==3.0.0",
    "attrs==25.1.0",
    "backoff==2.2.1",
    "bcrypt==4.2.1",
    "cachetools==5.5.0",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "comm==0.2.2",
    "copilotkit==0.1.35",
    "debugpy==1.8.11",
    "decorator==5.1.1",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "docopt==0.6.2",
    "email-validator==2.2.0",
    "executing==2.1.0",
    "fastapi==0.115.6",
    "fastapi-cli==0.0.7",
    "filetype==1.2.0",
    "frozenlist==1.5.0",
    "geoalchemy2==0.17.0",
    "google-ai-generativelanguage==0.6.10",
    "google-api-core==2.24.0",
    "google-api-python-client==2.159.0",
    "google-auth==2.37.0",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.3",
    "googleapis-common-protos==1.66.0",
    "googlemaps==4.10.0",
    "greenlet==3.1.1",
    "grpcio==1.70.0",
    "grpcio-status==1.69.0",
    "grpcio-tools==1.70.0",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "h2==4.1.0",
    "hpack==4.0.0",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "hyperframe==6.0.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==8.30.0",
    "jedi==0.19.2",
    "jinja2==3.1.4",
    "jiter==0.8.2",
    "jiwer==3.0.5",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "langchain==0.3.18",
    "langchain-anthropic==0.3.7",
    "langchain-core==0.3.34",
    "langchain-google-genai==2.0.9",
    "langchain-openai==0.3.4",
    "langchain-text-splitters==0.3.6",
    "langgraph==0.2.70",
    "langgraph-checkpoint==2.0.12",
    "langgraph-sdk==0.1.51",
    "langsmith==0.3.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mem0ai==0.1.48",
    "monotonic==1.6",
    "more-itertools==10.6.0",
    "msgpack==1.1.0",
    "multidict==6.1.0",
    "nest-asyncio==1.6.0",
    "openai==1.60.1",
    "orjson==3.10.15",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "partialjson==0.0.8",
    "passlib==1.7.4",
    "pexpect==4.9.0",
    "platformdirs==4.3.6",
    "polling2==0.5.0",
    "portalocker==2.10.1",
    "posthog==3.10.0",
    "prompt-toolkit==3.0.48",
    "propcache==0.2.1",
    "proto-plus==1.25.0",
    "protobuf==5.29.3",
    "psutil==6.1.0",
    "psycopg==3.2.3",
    "psycopg-binary==3.2.3",
    "psycopg2==2.9.10",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyannote-core==5.0.0",
    "pyannote-database==5.1.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pydantic==2.10.3",
    "pydantic-core==2.27.1",
    "pygments==2.18.0",
    "pyjwt==2.10.1",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "pyzmq==26.2.0",
    "qdrant-client==1.13.2",
    "rapidfuzz==3.11.0",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rich-toolkit==0.12.0",
    "rsa==4.9",
    "scipy==1.15.1",
    "setuptools==75.8.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "speechmatics-python==3.0.2",
    "sqlalchemy==2.0.36",
    "stack-data==0.6.3",
    "starlette==0.41.3",
    "tabulate==0.9.0",
    "tavily-python==0.5.0",
    "tenacity==8.2.3",
    "thefuzz==0.22.1",
    "tiktoken==0.8.0",
    "toml==0.10.2",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "uritemplate==4.1.1",
    "urllib3==2.2.3",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.3",
    "wave==0.0.2",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "websockets==14.1",
    "yarl==1.18.3",
    "zstandard==0.23.0",
]
